requirements.txt
streamlit>=1.32.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
langchain-text-splitters>=0.2.0

openai>=1.30.0

pymupdf>=1.23.8
pillow>=10.0.0
faiss-cpu>=1.8.0

tiktoken>=0.6.0
python-dotenv>=1.0.1
